src/tracebrain/requirements.txt
# TraceBrain Tracing Module Requirements
# Install with: pip install -r tracebrain/requirements.txt

# Core database ORM
sqlalchemy>=2.0.0

# PostgreSQL driver
psycopg2-binary>=2.9.0

# REST API framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# CLI framework
typer>=0.9.0

# HTTP client for tools
requests>=2.31.0

# SQL parsing for read-only enforcement
sqlparse>=0.5.5

# Vector search
pgvector>=0.2.5

# Optional: local embeddings (install only if EMBEDDING_PROVIDER=local)
# sentence-transformers>=2.7.0

# Optional: for SQLite JSON support enhancements
# (SQLite comes with Python, no installation needed)

# Optional: AI providers
google-generativeai>=0.3.0
openai>=1.0.0
anthropic>=0.30.0
